Type updates.

+++ Pulp/Common/INetwork.h
@@ -225,8 +225,8 @@ struct IPeer
 {
 	virtual ~IPeer() {}
 
-	virtual StartupResult::Enum init(uint32_t maxConnections, SocketDescriptor* pSocketDescriptors, 
-		uint32_t socketDescriptorCount) X_ABSTRACT;
+	virtual StartupResult::Enum init(int32_t maxConnections, SocketDescriptor* pSocketDescriptors,
+		size_t socketDescriptorCount) X_ABSTRACT;
 	virtual void shutdown(core::TimeVal blockDuration, uint8_t orderingChannel = 0,
 		PacketPriority::Enum disconnectionNotificationPriority = PacketPriority::Low) X_ABSTRACT;
